[Dev] Update and rename tests, add SVG test
