+++ src/lib/components/inbox/conversations/__tests__/zerostate_inbox-tests.tsx
@@ -1,3 +1,4 @@
+import "jest-snapshots-svg"
 import * as React from "react"
 import "react-native"
 import * as renderer from "react-test-renderer"
@@ -7,4 +8,9 @@ import Inbox from "../"
 it("looks correct when the user has no conversations", () => {
   const tree = renderer.create(<Inbox me={{ conversations: { edges: [] } }} />).toJSON()
   expect(tree).toMatchSnapshot()
+})
+
+it("renders correct when the user has no conversations", () => {
+  const tree = renderer.create(<Inbox me={{ conversations: { edges: [] } }} />).toJSON()
+  expect(tree).toMatchSVGSnapshot(320, 520)
 })
